backend/requirements-test.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
factory-boy==3.3.0
faker==20.0.3
pytest-env==1.1.3
pytest-xdist==3.3.1
pytest-mock==3.12.0
email-validator==2.1.1